// Frontend/src/app/classes/enums/document-type.ts
export enum DocumentType {
    CV = 'CV',
    CV2 = 'CV2',
    LETTRE_DE_MOTIVATION = 'LETTRE_DE_MOTIVATION',
    DEMANDE_DE_STAGE = 'DEMANDE_DE_STAGE',
    DEMANDE_DE_STAGE_SIGNER = 'DEMANDE_DE_STAGE_SIGNER',
    CLASSEMENT = 'CLASSEMENT',
    BULLETIN_DE_MOUVEMENT = 'BULLETIN_DE_MOUVEMENT',
    CONVOCATION = 'CONVOCATION',
    CONVOCATION_SIGNER = 'CONVOCATION_SIGNER',
    LAISSER_PASSER = 'LAISSER_PASSER',
    PRISE_DE_SERVICE = 'PRISE_DE_SERVICE',
    RAPPORT = 'RAPPORT',
    RAPPORT_SIGNE = 'RAPPORT_SIGNE',
    ATTESTATION = 'ATTESTATION'
}